simplify code

+++ monoio/src/driver/iocp/waker.rs
@@ -1,4 +1,5 @@
 use std::{io, sync::Arc};
+use std::ffi::{c_int, c_uint};
 
 use super::{CompletionPort, Event, Poller};
 
@@ -56,7 +57,7 @@ impl EventWaker {
         let buf = 0x1u64.to_ne_bytes();
         unsafe {
             // SAFETY: Writing number to eventfd is thread safe.
-            libc::write(self.raw, buf.as_ptr().cast(), buf.len());
+            libc::write(self.raw as usize as c_int, buf.as_ptr().cast(), buf.len() as c_uint);
             Ok(())
         }
     }
